Started adding tests for new TypeExtensions methods.
